Some changes in entity layer. Payments dao was implemented

+++ src/main/java/payments/dao/CardDao.java
@@ -6,5 +6,5 @@ import payments.model.entity.Card;
 import java.util.List;
 
 public interface CardDao extends CommonDao<Card> {
-    List<Card> findCardsByUser(int id);
+    List<Card> findCardsByUser(long id);
 }
