+++ src/main/java/payments/dao/CommonDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CommonDao<E extends BaseEntity> {
-    Optional<E> findById(int id);
+    Optional<E> findById(long id);
     List<E> findAll();
     void create(E e);
     void update(E e);
-    void delete(int id);
+    void delete(long id);
 }
